backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-sqlalchemy==3.1.1
python-dotenv==1.0.0
openai==1.51.0
google-generativeai==0.8.3
llama-cpp-python==0.2.82
pyjwt==2.8.0
werkzeug==3.0.1
